+++ changelog.md
@@ -1,5 +1,6 @@
 ### Next version
 * Upgraded logback from 1.5.6 to 1.5.16
+* Fixed issue in CompetingConsumerSubscriptionModel in which the model could be started automatically if consumption was granted, even though it was explicitly stopped
 
 ### 0.19.9 (2025-02-14)
 * Added toString() to subscription models for better debug output